.kiro/specs/dynamic-module-loader/requirements.txt
# 动态模块加载器 - Python 依赖包

# YAML 文件解析
PyYAML>=6.0

# 测试框架
pytest>=7.0.0
pytest-cov>=4.0.0

# 基于属性的测试框架
hypothesis>=6.0.0

# 代码质量工具（可选）
# pylint>=2.0.0
# black>=22.0.0
